Change money, accumulative, and quantity Int fields to Long to prevent overflow
Entities:
- Product.price, Product.popularity: Int → Long
- CartProduct.quantity, CartProduct.unitPrice: Int → Long
- Order.totalPrice: Int → Long

DTOs:
- ProductRequest (Create/Update).price: Int → Long
- ProductInfo/ProductResponse.price, .popularity: Int → Long
- AddCartItemRequest.quantity: Int → Long
- CartProductResponse.quantity, .unitPrice, CartResponse.totalPrice: Int → Long
- OrderResponse.totalPrice, RevenueResponse.totalRevenue,
  SpendingResponse.totalSpending: Int → Long

Service / Controller / Client:
- ProductService.incrementPopularity delta: Int → Long
- InternalProductController @RequestParam delta: Int → Long
- StoreServiceClient.incrementPopularity delta: Int → Long
- OrderRepositoryCustom + Impl: sumRevenue/sumSpending return Int → Long;
  fallback 0 → 0L
- StatisticsService.getRevenue/getSpending return Int → Long

Schema:
- products.price, products.popularity: INT → BIGINT
- cart_products.quantity, cart_products.unit_price: INT → BIGINT
- orders.total_price: INT → BIGINT

Tests: update all affected constructor calls and assertions

+++ order-service/src/main/kotlin/order/client/StoreServiceClient.kt
@@ -8,7 +8,7 @@ import org.springframework.web.client.RestClient
 @JsonIgnoreProperties(ignoreUnknown = true)
 data class RemoteProductInfo(
     val storeId: Long,
-    val price: Int,
+    val price: Long,
     val status: Boolean
 )
 
@@ -24,7 +24,7 @@ class StoreServiceClient(
             .retrieve()
             .body(RemoteProductInfo::class.java)!!
 
-    fun incrementPopularity(productId: Long, delta: Int) {
+    fun incrementPopularity(productId: Long, delta: Long) {
         restClient.put()
             .uri("$storeServiceUrl/internal/products/$productId/popularity?delta=$delta")
             .retrieve()
