+++ order-service/src/main/kotlin/order/dto/order/OrderResponse.kt
@@ -5,7 +5,7 @@ import order.entity.order.Order
 data class OrderResponse(
     val id: Long,
     val storeId: Long,
-    val totalPrice: Int,
+    val totalPrice: Long,
     val status: String,
     val createdAt: Long,
     val updatedAt: Long
@@ -26,11 +26,11 @@ data class RevenueResponse(
     val storeId: Long,
     val year: Int,
     val month: Int,
-    val totalRevenue: Int
+    val totalRevenue: Long
 )
 
 data class SpendingResponse(
     val year: Int,
     val month: Int,
-    val totalSpending: Int
+    val totalSpending: Long
 )
