+++ order-service/src/main/kotlin/order/repository/order/OrderRepositoryCustom.kt
@@ -3,6 +3,6 @@ package order.repository.order
 import java.time.ZoneId
 
 interface OrderRepositoryCustom {
-    fun sumRevenueByStoreAndMonth(storeId: Long, year: Int, month: Int, zoneId: ZoneId): Int
-    fun sumSpendingByUserAndMonth(userId: Long, year: Int, month: Int, zoneId: ZoneId): Int
+    fun sumRevenueByStoreAndMonth(storeId: Long, year: Int, month: Int, zoneId: ZoneId): Long
+    fun sumSpendingByUserAndMonth(userId: Long, year: Int, month: Int, zoneId: ZoneId): Long
 }
